Refactor `StagedDirectory` and `FallbackStagedDirectory`
This simplifies the functions implemented by `StagedDirectory` to two
methods: constructing an instance (with the side effect of staging a
directory) and the `captureAllOutputs()` to upload the files produced by
the execution of a command. It moves the rest of the methods to the
implementation of `FallbackStagedDirectory`.

It also introduces some changes in the `FallbackStagedDirectory`
implementation.