Add compile time option to select digest function
This adds the `BUILDBOXCOMMON_DIGEST_FUNCTION_VALUE` parameter, which
allows setting the digest function that the `CASHash` interface will
use.

It also adjusts some of the unit tests that previously were assuming
that hashes would always be generated with SHA-256 so that 100% of the
tests succeed no matter what digest function value is selected.