Switch metric enablement to be at MetricCollectorFactory level
Previously the feature to disable the capture of metrics was available both at
MetricPublisher scope as well as the individual metric guard scope. This
required to pass around the global 'is metrics enabled' state to every call
site that wanted to record a metric.

Now we simply ignore that flag for backwards compatibility (to later be
removed) and require the user to enable and disable only at the global factory
level.

It is not currently supported to adjust the enablement of metrics at runtime
since these methods are not thread-safe. In this change we are mostly concerned
with deprecating the old interface: the thead-safety aspect should be addressed
if any functionality is introduced to modify the metrics settings at runtime.

Signed-off-by: Ed Baunton <ebaunton1@bloomberg.net>
Signed-off-by: Marios Hadjimichael <mhadjimichae@bloomberg.net>