Introduce a ScopedMetric type to support holding a ptr to a metric
In order to provide support for scoping an existing metric, add a
`ScopedMetric` component that takes a pointer.

This moves most of the logic down from `MetricGuard` into a new component named
`ScopedMetric` such that `MetricGuard` simply becomes a holder of a copy of the
metric in question. `ScopedMetric` holds a pointer and takes care of publishing
the metric at destruction. `MetricGuard` must still exist so as to maintain the
interface that is currently in use.

Users can now opt to have a "fully managed" metric via the use of the
`MetricGuard` or construct their own metric and pass it into the `ScopedMetric`
which will publish the metric upon destruction but not destroy the metric.

This is required for implementing the counting metric: the user can construct
the counting metric outside of e.g. a loop, then pass it to the
`ScopedMetric`. They can continue to increment and operate on the existing
metric and know that it will be published when the `ScopedMetric` is
destroyed. This, instead of allowing users direct access to the metric held by
`MetricGuard`.

Signed-off-by: Ed Baunton <ebaunton1@bloomberg.net>