`Runner::executeAndStore()`: do not inline stdout/err
This removes the inlining in the `ActionResult` message of the contents
that a command outputs to stdout and stderr.

Those outputs, if not empty, will be uploaded to the CAS server and
their digests set in the `ActionMessage.std{out, err}_digest` fields.

Inlining will be the responsibility of an Action Cache.