CMake: Fetch googletest using CMake's FetchContent functionality
Newer versions of CMake provide functionality [1] to fetch external content and
resources that are not shipped with the source package.

Keep `GTEST_SOURCE_ROOT` CMake variable for passing in a local download of
googletest as a fallback as it was but otherwise fallback to fetching the
content from github. This will attempt to simply have CMake fetch googletest
from upstream at configure time. The simply reduces the complexity of using
googletest and googlemock and provides a simpler out of the box experience.

In keeping with the existing functionality, we first defer to the system wide
installs using find_package first; respect the user specified variable (if
passed); otherwise fall back to fetching the content remotely.

[1] https://cmake.org/cmake/help/latest/module/FetchContent.html

Signed-off-by: Ed Baunton <ebaunton1@bloomberg.net>