Runner: Drop directory deletion on signal
`d_outputPath` is a file, not a directory. It should also not be
necessary to delete the file as the process spawning `buildbox-run`
should delete it as part of its cleanup process. Otherwise, there would
be a race condition where the file doesn't get deleted if `SIGTERM` is
raised very late (after the last `getSignalStatus()` check).