Simplify gtest cmake search
Currently the search for the googletest modules is overly complex to try and
support a broad variety of versions. At this point, 1.8.1 is available in
debian stable and thus ships with the CMake confg files. This will work
most seamlessly.

Failing that we allow users to specify their own download of an appropriately
recent version of googltest using `GTEST_SOURCE_ROOT` CMake variable.

Finally, we simply attempt to grab the right version of googletest from the
internet and build it directly. Providing network connectivity is available,
this will also work fairly straightforwardly.

Signed-off-by: Ed Baunton <ebaunton1@bloomberg.net>