Export new target for gtest_main
Nearly every single user of this cmake module directly hardcoded `gtest_main`
in `target_link_libraries` which introduces a maintainence headache as it
requires that such a library be available by that name (which it implicitly is
because of the inclusion of this module). Rather than indirectly rely on the
availability of that library, provide a target that users can depend on to get
`gtest_main` in their tests.

Signed-off-by: Ed Baunton <ebaunton1@bloomberg.net>