`Client::upload()`: check `WriteResponse.committed_size`
When issuing a `ByteStream.Write()`, a server can abort the transfer
early if it detects that the digest is already present in its storage.

This makes sure that the `Client::upload()` method does not treat that
scenario as an error by checking that the
`WriteResponse.committed_size()` matches the size of the blob. If it
does it will consider the transfer a success even if the whole blob was
not transferred.

Updates the unit tests to set the previously-ignored `committed_size`
field in mocked calls.