Explicitly include libgflags on link-line
This is a workaround that includes libgflags for statically
linked applications. The shared library libglog.so contains
meta-data on it's dependencies which includes libgflags
in particular. As a result, link failures are not a problem when
linking dynamically, but it is when linking statically. Seeing
libgflags is included on the link-line indirectly for dynamically
linked applications, it shouldn't be harmful to explicitly include here.